fix: simplify missing ffmpeg handling for ASR
